implement code review changes
