+++ pages/plants/[id]/edit.js
@@ -1,4 +1,4 @@
-import PlantForm from "@/components/PlantForm";
+import PlantForm from "@/components/PlantForm/PlantForm";
 import { useRouter } from "next/router";
 
 export default function EditPlant({ plants, onEditPlant }) {
@@ -7,5 +7,5 @@ export default function EditPlant({ plants, onEditPlant }) {
 
   const plantEdit = plants.find((plant) => plant.id === id);
 
-  return <PlantForm $isEdit editPlant={plantEdit} onSubmit={onEditPlant} />;
+  return <PlantForm isEdit editPlant={plantEdit} onSubmit={onEditPlant} />;
 }
